backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
python-multipart>=0.0.9
pydantic>=2.8,<3.0
PyYAML>=6.0
Pillow>=10.0
httpx>=0.27,<1.0
torch>=2.2,<2.7
ultralytics>=8.4,<9.0
pandas>=2.2,<3.0
tqdm>=4.66,<5.0
seaborn>=0.13,<0.14
